viewset

+++ lule/apps/personas/api/views.py
@@ -13,10 +13,15 @@ from django.contrib.auth.models import User
 # Serializers
 from apps.personas.api.serializers import UserSerializer
 
+from apps.personas.api.permissions import UserPermission
+
 
 class UserListAPI(APIView):
 
+    permission_classes = (UserPermission, )
+
     def get(self, request):
+        self.check_permissions(request)
         paginator = PageNumberPagination()
         user = User.objects.all()
         paginator.paginate_queryset(user, request)
@@ -25,6 +30,7 @@ class UserListAPI(APIView):
         return paginator.get_paginated_response(serializer_data)
 
     def post(self, request):
+        self.check_permissions(request)
         serializers = UserSerializer(data=request.data)
         if serializers.is_valid():
             serializers.save()
@@ -35,13 +41,19 @@ class UserListAPI(APIView):
 
 class UserDetailAPI(APIView):
 
+    permission_classes = (UserPermission,)
+
     def get(self, request, pk):
+        self.check_permissions(request)
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request)
         serializers = UserSerializer(user)
         return Response(serializers.data)
 
     def put(self, request, pk, formatter=None):
+        self.check_permissions(request)
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request)
         serializer = UserSerializer(instance=user, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -50,7 +62,9 @@ class UserDetailAPI(APIView):
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
+        self.check_permissions(request)
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request)
         user.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
